Confirm closing of unsaved images (#6) on an image-by-image basis
I want to implement one final aspect of confirming close of unsaved
images - if the user has a bunch of unsaved images open, give them the
option to 1) save all, 2) prompt on an image-by-image basis, 3) ignore
all prompts.  Implementing this will take some thought, so I'll save it
for a future commit.