Added support for marking image forms as "internal"
Certain import routines - clipboard paste, screen capture - require a
temporary form for holding the imported image prior to processing.
CreateNewImageForm now allows calling routines to mark the created form
as "for Internal Use only", and when that form is unloaded the user will
(obviously) not be prompted to save the image in question.