Fix for #13 and #14, viewport problems related to maximized image windows
pdImages now tracks window state, and triggers a redraw of all viewports
when an image form is maximized or un-maximized.  (Necessary because MDI
rules require ALL child windows to be maximized or NO child windows to
be maximized.)  Also added checking for off-screen images; these are now
moved on-screen when an image form is un-maximized.