Histogram form is no longer modal
This means the user can leave the histogram form open as long as they
want.  When image windows are swapped (including when a new image is
loaded), the histogram is automatically redrawn.  Anything that passes
through the Process() routine also triggers a redraw.

The histogram is automatically unloaded when all child forms have been
unloaded.