Dynamically generate MDI child form icons (#9)
Custom MDI child form icons are now generated at form creation.  The
loaded image is resized to an aspect-preserved bitmap of max size 32x32,
converted to a transparent icon, then assigned to the form using
SendMessage.  (The traditional Form.Icon = ... doesn't work, because VB
will internally convert the icon to low bit-depth.)  All icon references
are stored in an array and destroyed at program termination.  The form
icons themselves are only 16x16, but I've chosen to render the icons at
32x32 in case it's ever decided to place them in the taskbar.  The size
can be changed easily via the line "icoSize = 32"; change that to 16, or
48, or whatever you wish.  It will automatically be resized on-the-fly
to 16x16 when assigned to the form.

I'm not closing #9 yet because the icon needs to be updated whenever the
image is changed.  This is not yet implemented.  As of this commit, the
form icon will always be the image as it appeared when initially loaded.