Fix unpredictable save behavior when closing the program with many unsaved images still open
If the user attempted to close PhotoDemon while having many unsaved
images open, it was possible for the images to be saved incorrectly due
to the way the active form was handled after each prompting.  This
behavior is now fixed, and attempting to close the program while many
unsaved images are still open will now lead to a proper cascading of
save prompts and, if applicable, saved files.