Undo/Redo buttons and menus now include descriptive text, Undo/Redo button redesign
Thanks to the implementation of GetNameOfProcess (see
abf0985b975bedb345e97fa147becde2ee2b9ed3) it is now possible to render
descriptive text to the Undo/Redo menus and buttons.

Also of note is that this commit includes a redesign of the Undo/Redo
buttons.  They have been condensed into image-only buttons, and their
tooltips are used to display the descriptive text of a given Undo/Redo
action.