Automatically assign system hand icon to clickable objects
I didn't think it was possible in VB6, but tonight I came up with a way
to enumerate all clickable forms in PhotoDemon and automatically force
them to use the system's hand icon.  I've wanted this feature for ages
because manually assigning a hand icon to all objects is a pain in the
ass at design time, and because VB doesn't support 32-bit cursors I've
been stuck using a 1-bit hand cursor from the Windows 95 era.

But all that has changed as of this commit.  PhotoDemon will now
automatically load the system's hand cursor (which is a beautiful
shadowed cursor on Windows 7) and assign it to all clickable objects on
a form.  Simply use "setHandCursorForAll Me" in the Form_Load event to
initiate.