Don't reset filter index upon common dialog cancellation
Steve M's common dialog class mistakenly (I presume?) resets the filter
index on common dialog cancellation.  Because this variable is passed
ByRef but is used elsewhere in the program, this was causing confusing
behavior in rare edge cases.  That value is now updated ONLY when the
common dialog is not canceled.