Added previewing to the grayscale conversion form
This was one of the final forms to be missing a preview function.  It
also uses a new system for previewing that doesn't require two functions
per filter (a preview and regular one).  Instead, it accepts a
"toPreview" flag and adjusts its run-time parameters accordingly.  This
dramatically cuts down on code maintenance, and it will need to be
mirrored across every other preview-generating form.   <sigh>