Allow grayscale reduction to two colors
The previous minimum was three.  I did this because I'd prefer that the
user user "Black and White Conversion" for converting to two colors, but
hey - if they want to use custom grayscale to do that, who am I to get
in their way?  (Also, the grayscale algorithm already handles two colors
so it's minimal work to enable this behavior.)