Automatically check for updated software versions
This commit only contains the code for CHECKING for updated software.
It attempts to access an online updates.txt file, extract version
numbers, then compare those numbers against the current app version.

If successful, it launches a blank form.  (Populating that form is next
on the to-do list.)  If unsuccessful, or if the software is up-to-date,
nothing happens.