Improved bug reporting capabilities on unknown errors
When PhotoDemon experiences an unknown error, the user is now asked to
submit a bug report.  They are then asked if they have a GitHub account;
if they do, they are sent to GitHub's "Submit New Issue" page.  If they
do not, they are sent to a PhotoDemon-specific contact page at
tannerhelland.com.  One final prompt with additional instructions is
displayed.