Began framework for applying filters to subsets of an image
As part of merging all core and preview filter/effect functions (#16)
I'll need to rewrite every filter and effect in the program.  (JOY!
/sarcasm)  While doing that, I'm also going to be dropping in a
framework that will ultimately allow functions to operate on just a
subset of the image.  Each filter will respect globally set (x, y) and
(width, height) values as determined by the GetImageData call.  At
first, those values will always be (0, 0) and (image width, image
height), but in the future, if a selection has been made, GetImageData
can modify the variables to the size of the selection and all
filters/effects will respect it.