Full implementation of previewing with DIBs
All a calling routine has to do is call prepImageData (and later,
finalizeImageData) with the Preview flag and the destination picture
box, and it takes care of the rest.  This means every PhotoDemon
function can be written with previewing support without any
modifications to the core filter code.  Mission accomplished.