Successful test of first grayscale routine with DIBs
Works flawlessly.  The net amount of code lines shouldn't change too
dramatically, based on what I observed with this function.  There will
be a bit more overhead thanks to the emphasis on local variables, but
this is offset by stripping out all preview detection work and moving
that into a single place.