Removal of Alfred's PCX/PNG/TGA classes
I appreciated Alfred's work, but his code is simply untested and
incredibly hard to debug (due to it being uncommente, unindented, and in
German).  GDI+ fulfills the PNG duties better with more reliability and
less overhead, and PCX/TGA are just not important to support in 2012.
(Also, FreeImage loads PCX and TGA well enough for me.)

This also meant the removal of the PCX Export form, but that's no big
loss.  I haven't tested that function with any rigor and as mentioned
above, no one should have need to export PCX files in 2012.