Display software logo dynamically (splash, about box)
This fixes several problems all at once.  There is now only one copy of
the PhotoDemon logo in the software (in a hidden picture box on
FormMain, because VB doesn't allow loading of JPEGs from resource
files...gofigure).  This logo is dynamically painted to the splash
screen and the About box as necessary, which not only saves on resources
- it also fixes #29 from Herman Liu.