Decouple image loading from UI (fix for #15)
Fix for #15.  Image loading is now completely independent of UI
considerations like creating a new MDI child form, rendering a viewport,
etc.  Simply call PreLoadImage with "isThisPrimaryImage" set to False -
and the optional parameters "targetImage" and "targetLayer", obviously -
and it will disable all UI functions and blindly load the file to that
image and/or layer.