Print preview is now a disaster - warning added
The feature needs to be rewritten against the new layer class, but the
last rewrite was such a complex clusterfuck that it's difficult to parse
out what actually needs to be fixed.  As such, it will probably required
yet ANOTHER rewrite, so I'm postponing that until after everything else
has been rewritten against layers.