Fixed viewport bug when switching from very zoomed-out to 100%
When extremely zoomed out on a smaller image, the scroll bars acquire
very large values.  If the user is scrolled over while zoomed in, then
suddenly switches to 100% zoom, the scrollbar value was not being reset
to zero and the image was being chopped off.

This is now handled correctly, and the scroll bar is reset to zero when
disabled.