Highly optimized viewport rendering
Some hacks were required to work around GDI obnoxiousness (like dropping
the alpha-channel on HALFTONE mode stretching) but it should be quite
fast now.  Selections can be drawn on a saved buffer that alleviates a
lot of the render strain.

Also, the checkerboard pattern on transparent images is now drawn
correctly.  Thank God.