Finished work on canvas drop shadow engine
It works quite well, very fast, and is only initialized once (thus
saving on precious resources and execution time).

Shadow strength and radius are currently set by constants in the code.
At some point these may be exposed to the user, but not today.

This is considered a superior replacement to "custom image borders"
described in #3, so I now consider that issue CLOSED.