New "FitWindowToViewport" method
Per Herman's excellent suggestion, when the "load images at 100% size"
preference is checked, the image window should not extend beyond the
boundary of the MDI client area.  This new routine is used to make sure
that image windows comfortably fit within that client area, and at
present it is only used when the image is NOT being "fit on screen"
automatically.