Added "remember last window location" preference
Note that this feature supports multiple monitors.  Sanity checks are
also performed to make sure the window doesn't appear completely
off-screen. (For example: if PhotoDemon is used on a secondary monitor,
then closed, and that monitor is detached, the next time PhotoDemon is
run it will automatically return itself to the primary monitor.)