Allow the user to change an image's color depth
There are now corresponding menu entries for 24 and 32bpp mode, via
Image -> Mode.  The top-level icon for "Mode" handily changes to reflect
the current image mode.  Checkmarks are rendered quite beautifully on
Vista/7 if theming is enabled; otherwise, custom menu icons are used.

Converting a 32bpp image to 24bpp results in a composite against white.
In the future, the user may be given more choices for the downsampling.