Calculate and remember initial image color depth
This is done for all images, regardless of source or import method, via
a custom color-counting solution.  Recognized values are 1, 4, 8, 24, 32
bpp.  Note that the image data is still stored in 24bpp or 32bpp format,
per PhotoDemon requirements - but the relevant pdImage will remember the
initial BPP value of the image.  This can be used later, upon image
export, to write out the image at its initial bitdepth.