Disable dynamic Recent File icons on XP
So this is somewhat convoluted, but - I currently use two separate icon
managers: one for the program's static icons, and another for the
dynamically generated Recent File icons.  This is necessary because the
Recent File icons undergo a lot of churn during program execution (as
they are moved around, saved, loaded and re-loaded) and I need to
frequently clear the whole thing out - something that obviously cannot
happen to the main icon cache.

The problem is that in XP, I have to subclass the menu and owner-draw it
in order to show icons.... but two menu icon classes means dual
subclassing, which is a major no-no.  Rather than rewrite the entire MRU
icon handler to operate on a single handler class (which would be a pain
in the ass), I'm disabling MRU icons in XP for now.

Another problem is that XP doesn't allow sub-menus to use different icon
dimensions than the parent menu, which means no 64x64 icons for recent
files.  16x16 isn't nearly as useful, so I'm not in a huge hurry to make
this work on XP.

That said, I do plan to revisit this at some point in the future, but
for now I need to focus on wrapping up the few final things needed to
push 5.2 out the door.