Stricly enforce "Cancel" behavior of Preferences "Cancel" button
Before this commit, there were a few preferences (such as Interface
Text) that would not be properly restored if "Cancel" was pressed after
a change was made - this was owing to their nature of "real-time
updating" preferences.

Now, however, the original setting is restored upon cancel for any such
preferences.

(This was a blocker for #41)