Vastly improve MRU hash performance by storing per-session hashes
Previously, every time PhotoDemon needed to access an MRU icon, it would
request a hash of the MRU path in question.  As SHA-256 hashing is
computationally expensive, this led to a lot of wasted cycles.

With this update, every hash from the current session is remembered.
When a hash of a file path is requested, the MRU handler first checks
its list of hash requests processed this session.  If a path has
previously been requested, the last calculated hash is returned (instead
of calculating it again).

Much better.