Remove rare J2K format extensions to fix filelistbox.pattern crash
VB has a limit on the size of strings passed to filelistbox.pattern (who
knew?) and PhotoDemon exceeds that limit when all its file formats are
enabled.

Pain in the ass to fix - but for now, the workaround is to simply remove
some rare J2K extensions to put us back under the limit.