Redesigned Preferences dialog, "Performance" added, Load/Save separated
Looking at the roadmap for 5.4, I realized there are a lot of new
under-the-hood features that should be exposed to the user.  In lieu of
a better arrangement, I'm planning on sticking these into the
Preferences dialog.

Load/Save needed to be separated because there will be a lot of
preferences per file format (especially for saving).  The combined
dialog page from 5.2 was already full.

Performance will be used for the new swap-to-file and compression
behavior being added.  I considered putting these under "Advanced", but
they're not advanced, and they're going to be completely safe to use -
so no, I don't think they belong there.  A dedicated section is better.