Implement "Close All" command
The behavior of this command is surprisingly sophisticated - take, for
example, the user having 10 open images, all of which are unsaved.  If
they select "Close All", the program will attempt to close the
first-loaded image, realize that it is unsaved, and thus display a
confirmation prompt.

If, on this prompt, the user selects "Save" and "Repeat for all images",
the program will prompt the user for a save file name, and if necessary,
any additional save options (such as JPEG quality).

Say this goes fine for the first 5 images, but then on image 6 the user
realizes he doesn't actually want to close the image - so he hits
"Cancel" on the common dialog form.  This will prevent that image from
being saved.

It will also switch the program back into "confirm each image
independently" mode, so that the user can manually specify save vs do
not save vs cancel behavior for any remaining images.

Very nice.