New and much-improved incrementFilename function
When a file is saved via Save As to a directory and a file with that
name already exists, the Save As function is smart enough to increment
the filename with the format "filename (n+1)".  If that file is also
found, the function continues until a unique filename is found.

The problem with the old function was if you opened a file with a name
like "filename (2).jpg", a Save As would attempt to save the file as
"filename (2) (2).jpg".  The new function is smart enough to recognize
the trailing number and stick to the "filename (n+1).jpg" formula.

This will also be important for a new preference regarding "Save" -
whether it is supposed to overwrite the original file, or simply save a
copy.