New preference for "safe" save behavior
When using the "Save" function, standard behavior is to overwrite the
old file.

PhotoDemon now supports "safe" save behavior, where the original file is
preserved, and the current file is saved as "filename (n+1).extension".
Inside the program, the active window caption still displays the
original filename, but the MRU is updated with each revision as they are
saved.

The standard behavior is enabled by default to avoid confusion.