Make FitWindowToImage way more sophisticated
If fitting the window around the image will result in part of the image
falling off-screen, move the window around and resize the borders as
necessary.

Note that this isn't a perfect solution, as the presence of scroll bars
may lead to a window size that's slightly too small (all scroll bar
handling is handled in the separate viewport handler code).  I may have
a way around this, but I will save it for a separate commit.