Allow user to hide/show the left-hand panel
The View menu now contains an entry for "Show Left Pane" or "Hide Left
Pane", with an accompanying image that changes dependent on the current
state of the pane.  The pane's presence is saved to the INI file as
well.

This was done in preparation for a right-hand "tools" pane, which will
also need to be dynamically visible.