Manage support image memory usage much more aggressively
PhotoDemon generates a LOT of icons.  Typically these are stored in a
persistent cache that is cleared only at unload time.  Unfortunately,
this strategy is probably not ideal on low-end systems, especially as
the number of icons has gone steadily up as new features are added.

With this patch, any time the number of loaded images equals zero, the
icon cache is cleared and all associated memory is releases.  It doesn't
make a massive difference, per se, but when working with a ton of images
it does help keep memory usage down.