Completely overhauled menu icon handler
Menu icons are now loaded and placed by a single call.  This is much
more organized than the previous (default) method, and it has the
ENORMOUS benefit of not requiring me to manually track the ID of every
loaded icon.

It also means I could finally clear out a number of old, unused icons.