Implement new subclassing technique for form theming
Themed controls do not render their background properly.  This subclass
fix forces transparent rendering of all control backgrounds, which is
exactly what we want.

Note that it is wrapped within a new "ReleaseFormTheming" sub that is
called at the unload of every form.

This feature is disabled in the IDE to prevent crashes upon pressing the
"Stop" button.