Restore form backcolor
Note to self: the interface of various forms is potentially going to be
wacky these next few weeks as I work on a number of interface changes.
I'll try to avoid mixing interface adjustments into other commits but
given my track record, it's unlikely to be a flawless transition.  :p