Interface overhaul: step 1
All forms now use a copy-and-pasted command button setup for their
OK/Cancel buttons.  This ensures several things:

1) All OK/Cancel buttons are the same size.
2) OK and Cancel buttons all have accelerators set up.
3) OK is always set as "Default" button, and "Cancel" is always set as
"Cancel" (so Esc can be used to trigger it)
4) I have also added a small bit of coloring to distinguish the command
button area from the rest of the forms - important, because:
5) All forms now use the system "window background" color (instead of
previous "button face").  On Windows 7, this means a white background.
Much more modern.