Epic interface overhaul step 2: previewing
All effect tools with live previews have been redesigned in this commit.

Instead of using plain picture boxes, effect previews now operate on a
custom-built "fxPreview" control, which I have created specifically for
PhotoDemon.  This allows me to add new preview features - for example,
toggling between the preview and the original image - and have the
feature automatically available to every tool in the project.

The presence of a single preview box (instead of a "before" and "after",
as before), also provided a good opportunity to redesign all tool forms
to a new, more modern layout.

Tools are now horizontally distributed, with a large preview on the
left, and tools options on the right.  This is a better fit for
widescreen monitors (and dual-monitor layouts), and it also means the
program can *theoretically* be run at an 800x600 resolution.  I don't
actually care about this, but I thought it was a nice benefit.  :)

All effects and tools have been checked to make sure they can render the
new preview size in real-time.  All can, though some will be stuttery in
the IDE.

A number of other, smaller interface tweaks and bugfixes were added as
part of this commit, but I don't plan on breaking them out separately.
(These were discovered while testing all tools on the new preview
system.)