Heavy-duty gaussian blur optimizations
The speed savings of this commit vary; by my calculations it looks to be
anywhere from 10-30%, with higher savings for smaller images.
Optimizations include:

1) Massive lookup table for gaussian values; this is where the bulk of
the savings occur, as it eliminates a large amount of floating-point
multiplication from the inner loop.

2) Hackish use of 1D arrays to access source and destination pixels on
the first pass, and destination pixels on the second pass.  VB takes a
big speed hit with multidimensional arrays; by creating a SafeArray
header that points to each line, and updating that with each new line,
we can fool it into allowing the use of 1D arrays to access 2D data.
The savings aren't huge, but it's enough to shave a couple seconds off
large blurs of photo-sized images.

I've hit a wall with a few other ideas that aren't working out, but I
hope to be able to shrink the time further in the future.  (This is
especially important as a number of other filters rely on the central
gaussian blur function.)