New lens distortion correction tool
In the Effects -> Distort menu, I have added a new tool for correcting
lens distortion.  This is a pretty neat ability, and I have coded it
openly so the user can fix any lens, regardless of original settings.
(Some other tools rely on knowing the original focal length.)

The trade-off is that the user must really play with the scroll bars to
get the perfect correction, and optionally apply a crop afterwards, but
I think that's a fair trade-off in this case.