Validate language files, including basic XML parsing
At startup, the /Data/Languages/ subdirectory is now searched for
language files.

Files are partially loaded (first 1024 bytes), and validated using a
series of criteria.

Valid files are then further parsed, with key information being pulled
from the language file, including:

Language ID (e.g. "en-us") - note that this is the only required tag for
the language file to function
Name (e.g. "English (United States)")
Version (e.g. "1.2.3")
Status (e.g. "Complete", or "Incomplete", or "Needs work" - anything can
be placed here, as long as it's descriptive)

The basic XML parser I'm implementing isn't really a parser, per se.
I'm not interested in building a full schema or grammar handler.  This
is more of a "fast and dirty" approach to retrieving information, and
the XML format is primarily used for just that - formatting.

That said, the first tool I've added - a simple function for returning
the text for a given tag - is fast and elegant, which is more than good
enough for me.