Fix translations for custom shortcut menu entries
Because the custom shortcuts are added at run-time, they are returned as
part of the menu caption.  Thus when trying to retrieve the translation
that menu, something like "   Ctrl+X" is appended to the end.

Now, custom shortcuts are stripped out before translations are applied
(and the shortcuts are restored afterward).