Add right-hand tool pane
I've been meaning to do this for a long time, as the selection tool
options don't fit well on the right, and more tools are forthcoming.
Now seemed like a good time.

Translation support requires some modifications to the layout on the
left pane, so I needed to implement this before a full language support
implementation.