Finish adding translation support to all modules and classes
This is a big update, and from this point forward any missing
translation code is a genuine mistake.  I have tried to be as thorough
as possible, but when you have hundreds of files to manually inspect,
some mistakes are possible.

The next step is to somehow generate a master translation file.  Parts
of it can probably be automated, but some manual inspecting will also be
necessary.