New class for adding PNGs to controls at run-time
Now it will be possible to assign 32bpp images to any hWnd.  First, I'll
be using this to phase out some instances of jcButton where it's no
longer needed.

The only problem with this technique is that the images look like shit
when the parent control is disabled (and I'm not sure why), so for now
it should only be used on controls that are guaranteed to be enabled for
their entire lifetime.