Start laying groundwork for better "first-run" experience
Namely, prompt the user for a language rather than throwing one at them
and forcing them to hunt for the Language menu.  (Since PhotoDemon
doesn't have an installer, first-run is the earliest we can ask for such
a thing.)

In the future, I may also present the user with a choice of visual
theme.  This can be helpful for disabled users.