Further Gaussian Blur optimizations
I was able to cut a further 10% of processing time with another set of
carefully profiled optimizations.  Classic VB is painfully difficult to
optimize on account of its total unpredictability.  So many unrelated
factors affect performance, and looking at the code you'll notice that
some optimizations work in the x-direction but not y (and vice versa).
Strange stuff.

About the only thing left I can do to speed up the function is write a
new Integer-only version.  I'm not sure how worthwhile that would be,
but I may investigate in the future after some other projects wrap up.