Updated German language file
Hand-edited to fix some lingering issues and a few missing translations
that I am capable of replacing (mostly by mimicking the German text of
other photo editing tools).

Should be up-to-date minus the Batch Wizard, but there is enough text
from the old Batch Convert form that I think a lot of the text can be
successfully recreated.