Language file update
For performance reasons, I'm trying a new system with language files.
Duplicates are no longer omitted from the final language files.  This
allows me to search within a form's specific area of the translation
file for text, which cuts down the time searching for strings (not as
much as you'd think, but every little bit helps).