Add new edge pixel handler "ignore" to all Distort tools
Thanks to the brilliant original implementation, it was all of 10 lines
of code to add this to every Distort tool - I just added a new constant
to the central handler, then added a few lines of code to the
pdFilterSupport class.  And BAM, every Distort tool gets a new edge
handler.  Yay!

This allows for identical output to Robert's PaintRR, and also older
versions of Paint Shop Pro.