New custom control that merges a text box and scroll bar as an integrated text/up-down
The traditional up/down control requires an added typelib that I have no
use for... so I built my own.  My control has the added bonus of
integrated max/min, floating point support, a single "Change" method,
and a single "isValid" method, complete with a red error box for invalid
entries.

It's quite nice, and it saves me a lot of code in the long run.  It also
solves the localization problem of decimal vs comma for floating-point
values.