Allow selections to begin anywhere on viewport (fix for #51)
As stated in the title, this commit fixes Issue #51 - "Cannot begin
selection outside of image".  You can now begin a selection anywhere on
the viewport, and once the mouse is moved over the image, the selection
will begin as expected.  This was a great usability suggestion from
Kroc.

Unfortunately, this commit breaks "move selection" behavior, and I'm not
sure why - it may take some work to track down a fix, but I wanted to
have this solution in its own commit in case I need to revert back.