Provide interface controls for antialiasing and feathering
NOTE: feathering isn't implemented yet (all the plumbing is there, but a
feathering pass is never applied).  But antialiasing works brilliantly,
and all options match up between non-antialiased and fully AA.

Note that with this implementation, feathering will require the
existence of AA.  That is by design.  It makes no sense to feather a
non-AA selection as the results will always be subpar.