Add preliminary support for selection feathering
Technically, selection feathering works with this commit.  But some
obvious problems exist... some of which I can fix, some of which I might
not be able to.

The biggest problem is performance.  On a large selection, gaussian blur
is not cheap.  It is possible to use GDI+ to render the blur - I may
examine that to see if it's a viable replacement.

The easier fix I can make is not re-rendering the mask before every
action.  I'll be adding some tracking variables to the selection class
to only re-render the mask when absolutely necessary.