Temporary commit with all my export metadata work
If FreeImage ever decides to offer proper metadata exporting, this code
would allow me to easily make use of it.  But in the meantime, I am
forced to investigate other options - so after this commit, all this
code is getting deleted.

But I wanted to make sure I archived it here.  :)