Properly handle binary metadata
All binary data - except thumbnails, which PhotoDemon will generate
itself - will now be ignored by the tag parser.  The tag itself will
still be parsed and named, but rather than extract out the binary data
(which we can't show to the user anyway), just mark the tag as "binary"
and leave it at that.