Fix long-standing tooltip theming bug (again)
This bug just will not die.  I've now discovered that VB doesn't like
you passing a non-instantiated object ByRef, then attempting to
instantiate it in the receiving function.  It won't throw errors, but
the object seems to become lost in space and time.

The solution seems to be forcibly instantiating the object *before*
passing it ByRef.  I hope this is the last time I have to revisit this
bug.