Improving selection feathering performance on XP/Vista
As with the GDI+ fix submitted in a prior commit, this restricts the
feathered (e.g. gaussian blurred) area to the smallest possible region
of the selection mask.  There is no performance differences for
image-sized selections (as the full mask still needs to be blurred), but
for smaller selections, this should provide a welcome acceleration.