Convert "Pan and Zoom" zoom scale to [-10, 10]
The programmer in me hates to see a zoom value of "0", but honestly this
seems much more intuitive to a casual user.  "0" means no change,
negative values mean zoomed out, and positive values mean zoomed in.
Easy enough.  Previously, the scale of 0.01 to 64 was terrible because
it meant only 1/64 of the scroll bar was dedicated to zooming out.