New class for detailed font and text management and rendering
Eventually I'm going to have to implement some kind of text tool, so I
thought I'd start laying the foundation for it.  Thus this text and font
management class.

At present, the class's emphasizes providing a simple way to manage font
attributes, create an API font, select that font into a DC (with color
and alignment), then do high-performance rendering to that DC.

For sample output and usage, see the About form.