Rewrite About dialog against new pdFont class
The new "credit scroll" aspect of the About dialog provided a great test
bed for the new pdFont handler.  Performance is much improved over pure
VB, probably because VB has to recreate a font object every time a
single parameter is changed.

Setting up and using the font object is very simple, and the About
dialog LOC was cut by some 25% thanks to this new approach.