Add PhotoDemon's master language creation tool to repository
This tool is used to generate a master English-language XML file of all
user-facing text in PhotoDemon.  It is not intended for any other use,
and I have no intention of ever turning it into a general-purpose tool.

I am adding it to the main PD repository because it is an indispensable
part of the project, and I have to update it with some frequency - so I
would like changes tracked.