Re-order batch wizard pages and improve wizard text
Starting with the "select images" panel was probably a bad idea, as it
is by far the most daunting of the panels.  Now, I start with "what
action(s) do you want to apply to images?", which is much friendlier.

The wizard also describes what a batch process is, in case someone
doesn't know and is just exploring the tool.