New AlphaBlend API support function for layers
Useful when rendering a layer directly to a DC with full alpha support.
Premultiplication is required, and layers do not currently offer a
premultiplication function, so the assumption is that the layer has been
directly created from an image in PD's resource section via GDI+ (so
premultiplied alpha is already in effect).