Selections can now lie off the image (persistently)
See comments at top of pdSelection for more detail; basically, bounds
checking for selections is now greatly relaxed.  The selection can lie
anywhere off the image, so long as at least one pixel is in bounds.
This is helpful with elliptical selections as you can now select
arc-shaped portions of the image.