Fix Canvas Size anchor tool appearance in XP
In this latest round of my least favorite game ("Fixing bugs for
obsolete OSes!") I discovered that XP has serious issues changing a
button's image once it has been loaded.  I could switch to an
owner-drawn approach to fix this, but it would make the buttons look
significantly worse in Vista/7/8.

I'm not going to do that, so XP users are stuck with an uglier,
non-image version of the anchor tool.  Sorry!