New "Line Selection" tool
I've found this to be a surprisingly helpful tool, and one that is not
easily replicated in either GIMP or Photoshop.  Sometimes one wants to
select a line-shaped region, and this new dedicated tool makes that
simple.

One application I've looked at heavily is a tilt-shift feature.  I'll
need to improve possible feathering radii for it to work, but I think
the result could be quite striking.