Fix saved selection files to use RELATIVE values
What makes this so great is that saved selections will correctly appear
over any size of image.  For example, my test selection is a line from
the image's bottom-left to top-right corners.  The size of the image
doesn't matter, because values are saved on a [0.0,1.0] relative scale,
then converted when the selection file is loaded.  This means that you
can load any image of any size/aspect ratio, and the saved selection
file will always result in a line from the bottom-left to top-right
corner.