Main menu overhaul
This update brings the main menus into closer parity with Photoshop.
The goal was not to "copy Photoshop", per se, but to more intelligently
organize various options - and it just so happened that this resulted in
a menu closer to Photoshop's, but with some key differences.

Items of note:
- Color menu is now "Adjustments".  It contains operations helpful for
technical editing of photos.
- The new Adjustments menu is subdivided into three primary blocks:
--- Operations that affect lighting
--- Operations that affect color balance
--- Operations that apply wholesale composition changes (e.g. grayscale,
sepia)
- Some color menu items have been moved to the Image menu because they
are either metadata-based (count colors) or because they cannot be
applied to selections (Indexed color, previously Reduce color count)
- Some color menu items have been moved to the Effects menu, because
they were purely stylistic.  This resulted in minor changes to some
Effects menus.
- Some color menu items have been removed entirely (all Enhance options,
all Fade options).  These options may return in a future update pending
a rewrite to make them more useful.  The Enhance items will most likely
be rewritten as part of an "Auto" menu, which will include items like
"Auto Levels".